Robust runner/reporter support
Instead of patching in support for sending test data over the websocket
from the playwright script, use the existing js-reporters library to do
this properly.

However, this requires either copy/pasting js-reporters into the
playwright script or making the addon consumer run this setup on their
own.

Having consumers run the setup on their own is the better option of the
two, imo.